fix: threading
