+++ MVVMaApp/MVVMApp/View Controllers/Book Details/View Controller/BookDetailViewController.swift
@@ -58,7 +58,7 @@ class BookDetailViewController: UIViewController {
     private func configure(viewModelGetObject: BookDetailViewModel?) {
         viewModelObject = viewModelGetObject
         guard let thumbnail = cellViewModel?.thumbnail else { return }
-        viewModelGetObject?.loadImage(url: thumbnail) { (image) in
+        ImageLoader.loadImage(url: thumbnail) { (image) in
             DispatchQueue.main.async {
                 self.bookImage.image = image
             }
